Move configuration files under src and update references
