+++ src/Lotgd/Config/configuration.php
