+++ src/Lotgd/Config/configuration_extended.php
