+++ src/Lotgd/Config/settings_extended.php
