+++ tests/bootstrap.php
@@ -1,7 +1,7 @@
 <?php
 
 require __DIR__ . '/../autoload.php';
-require __DIR__ . '/../config/constants.php';
+require __DIR__ . '/../src/Lotgd/Config/constants.php';
 
 require_once __DIR__ . '/Stubs/DbMysqli.php';
 require_once __DIR__ . '/Stubs/Database.php';
